sensors interface
